Writing output to CSV is now possible.
DHtmlWriter:
writeNodeAllMetricsAsVariousFormats now splits into CSV and Json.
The CSV file now has the name of the file with the snapshot.

MetricData:
Renamed from CSVData and moved to Code-Analysis export metrics
SnapshotDates are in the csv file as SimpleDateFormat and are now displayed as first entry in all csv files.

MetricDecorator:
Renamed from CSVDecorator and moved to Code-Analysis export metrics